trace: Re-order trace creation
If we had any un-cache trace classes, we currently call finalize_objects, which
calls the finalize_object method even for trace classes that _were_ cached. For
those cached trace classes, we had already called finalize_object before
creating the cache, so any operations in finalize_object will now be done
twice. All the current finalize_object methods seem to be idempotent so this
isn't a problem, but is nice to avoid. Therefore just create the cache before
calling finalize_objects.

More importantly, create the cache before calling normalize_time. This means we
cache CSV with time un-normalized and normalize it when needed, so that the
normalize_time parameter to the Trace object will not be affected by caching.